backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "smartshelf-backend"
version = "0.1.0"
description = "Backend API for SmartShelf inventory management"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "SmartShelf Team"}
]
dependencies = [
    "fastapi>=0.95.0",
    "uvicorn>=0.22.0",
    "pydantic>=2.0.0",
    "pandas>=2.0.0",
    "joblib>=1.3.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.12.0",
    "python-dotenv>=1.0.0",
    "openai>=1.5.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0"
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I", "N", "W", "B", "A", "C4", "UP", "PT"]

[tool.ruff.isort]
known-third-party = ["fastapi", "pydantic", "uvicorn", "sqlalchemy"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[project.urls]
Documentation = "https://github.com/yourusername/smartshelf#readme"
Source = "https://github.com/yourusername/smartshelf"
"Bug Tracker" = "https://github.com/yourusername/smartshelf/issues"

[tool.hatch.build.targets.wheel]
packages = ["app"]
